feat(ui): expose the client-side form field frame
Move the pure label/helper/tooltip/error frame into @lattice-php/ui as
FormField with a plain bare prop; the form package's FormFieldFrame now
only resolves bare from the table-cell context so its public API stays
unchanged.
